# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Diasoft Platform API
  version: 1.2.0
  description: Shared OpenAPI contract for the Diasoft diploma verification platform. The file is stored in diasoft-gateway and documents public gateway routes, gateway private auth routes for the new frontend, and registry backoffice routes used by the current diasoft-web.
servers:
  - url: http://localhost:8080
    description: Gateway local runtime
  - url: http://localhost:8081
    description: Registry local runtime
  - url: https://verify.example.com
    description: Public gateway runtime
  - url: https://registry.example.com
    description: Internal registry runtime
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ImportId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    DiplomaId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ImportJobError:
      type: object
      required: [id, importJobId, rowNumber, code, message, createdAt]
      properties:
        id:
          type: string
          format: uuid
        importJobId:
          type: string
          format: uuid
        rowNumber:
          type: integer
        code:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    Diploma:
      type: object
      required: [id, universityId, universityCode, studentId, studentExternalId, ownerNameMask, diplomaNumber, programName, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        universityId:
          type: string
          format: uuid
        universityCode:
          type: string
        studentId:
          type: string
          format: uuid
        studentExternalId:
          type: string
        ownerName:
          type: string
          nullable: true
        ownerNameMask:
          type: string
        diplomaNumber:
          type: string
        programName:
          type: string
        graduationYear:
          type: integer
          nullable: true
        recordHash:
          type: string
          nullable: true
        status:
          type: string
        verificationToken:
          type: string
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        revokeReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RevokeRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    ShareLinkRequest:
      type: object
      properties:
        maxViews:
          type: integer
          minimum: 1
    ShareLink:
      type: object
      required: [id, diplomaId, shareToken, expiresAt, usedViews, status, shareUrl]
      properties:
        id:
          type: string
          format: uuid
        diplomaId:
          type: string
          format: uuid
        shareToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        maxViews:
          type: integer
          nullable: true
        usedViews:
          type: integer
        status:
          type: string
        shareUrl:
          type: string
paths:
...
  /api/v1/imports/{id}/errors:
    get:
      tags: [registry-imports]
      summary: Get import row errors
      operationId: getImportErrors
      parameters:
        - $ref: '#/components/parameters/ImportId'
      responses:
        '200':
          description: Import errors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImportJobError'
...
  /api/v1/university/diplomas/{id}/revoke:
    post:
      tags: [registry-university]
      summary: Revoke a diploma
      operationId: revokeDiploma
      parameters:
        - $ref: '#/components/parameters/DiplomaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeRequest'
      responses:
        '200':
          description: Revoked diploma
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Diploma'
...
  /api/v1/student/diplomas/{id}:
    get:
      tags: [registry-student]
      summary: Get a student-scoped diploma
      operationId: getStudentDiploma
      parameters:
        - $ref: '#/components/parameters/DiplomaId'
      responses:
        '200':
          description: Diploma detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Diploma'
  /api/v1/student/diplomas/{id}/share-links:
    post:
      tags: [registry-student]
      summary: Create a share link for the current student scope
      operationId: createShareLink
      parameters:
        - $ref: '#/components/parameters/DiplomaId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareLinkRequest'
      responses:
        '200':
          description: Share link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLink'